build: add header files dependencies
Since we are now building greybus in kernel tree and headers are not
available under the regular uapi for user space to use them, let's add
now the greybus_protocols and greybus_manifest header files to gbsim and
try to make them in sync with any kernel change if needed it.

Remove the entry in the Makefile and configure regarding the need for a
GBDIR specific variable directory.

Signed-off-by: Rui Miguel Silva <rmfrfs@gmail.com>