functionfs: add arpc handler
To be able to reply to arpc messages we need to add the arpc header
file, and also handle the arpc run with a request attached.

We use the previously introduces to_ap_arpc endpoint to send a response
to AP. Right now we just reply success to any command, that may change
in the future.

Signed-off-by: Rui Miguel Silva <rmfrfs@gmail.com>